telegram/requirements.txt
nats-py==2.11.0
requests==2.32.5
ujson==5.11.0